feat(main-ui): add MyPageLayout and refactor IndexPage for improved structure
- Introduced `MyPageLayout` component to enhance layout flexibility.
- Refactored `IndexPage` to utilize `use_context` for `Controller` and simplified its structure.
- Updated `Route` to use `MyPageLayout` for the `/social` route.
- Enhanced modularity by reorganizing imports and reducing redundancy in layout components.
