+++ packages/main-ui/src/pages/page.rs
@@ -1,136 +1,38 @@
-#![allow(unused)]
-use crate::{
-    components::loader::Loader,
-    pages::components::{BottomSheet, CreateFeedBox, FeedContents, LeftSidebar, RightSidebar},
-};
+use crate::pages::components::FeedContents;
 
 use super::*;
 use bdk::prelude::*;
 use controller::*;
-use dto::ContentType;
 use i18n::*;
 
 #[component]
 pub fn IndexPage(#[props(default = Language::En)] lang: Language) -> Element {
     let mut is_write = use_signal(|| false);
-    let mut ctrl = Controller::new(lang)?;
+    let ctrl: Controller = use_context();
     let tr: IndexTranslate = translate(&lang);
 
     let landing_data = ctrl.landing_data()?;
-    let hot_promotions = ctrl.hot_promotions()?;
-    let news = ctrl.news()?;
     let profile = ctrl.profile()?;
-    let communities = ctrl.communities()?;
-    let accounts = ctrl.accounts()?;
 
     let my_spaces = landing_data.my_spaces;
     let following_spaces = landing_data.following_spaces;
 
-    let followers = landing_data.follower_list;
-    let profile_data = landing_data.profile_data;
-
-    let recent_feeds: Vec<String> = my_spaces
-        .iter()
-        .map(|v| v.title.clone().unwrap_or_default())
-        .take(3)
-        .collect();
-    let recent_spaces: Vec<String> = my_spaces
-        .iter()
-        .map(|v| v.title.clone().unwrap_or_default())
-        .take(3)
-        .collect();
-    let recent_communities: Vec<String> = communities
-        .iter()
-        .map(|v| v.title.clone().unwrap_or_default())
-        .take(3)
-        .collect();
-
     rsx! {
         by_components::meta::MetaPage { title: tr.title }
 
-        div { class: "flex flex-col w-full h-screen justify-start items-start",
-            div { class: "flex flex-row w-full h-[calc(100vh)] justify-start items-start py-20 px-10 gap-20",
-                LeftSidebar {
-                    lang,
-                    profile: profile_data.clone(),
-                    // recent_feeds: recent_feeds.clone(),
-                    // recent_spaces: recent_spaces.clone(),
-                    // recent_communities: recent_communities.clone(),
-                    accounts: accounts.clone(),
+        FeedContents {
+            lang,
+            my_spaces,
+            following_spaces,
+            profile: profile.profile.clone(),
 
-                    onwrite: move |_| {
-                        is_write.set(true);
-                    },
-                    add_account: move |_| async move {
-                        ctrl.add_account().await;
-                    },
-                    sign_out: move |_| async move {
-                        ctrl.signout().await;
-                    },
-                }
-
-                FeedContents {
-                    lang,
-                    my_spaces,
-                    following_spaces,
-                    profile: profile.profile.clone(),
-
-                    is_write: is_write(),
-                    onwrite: move |_| {
-                        is_write.set(true);
-                    },
-                    onclick: move |id: i64| {
-                        ctrl.move_to_threads(id);
-                    },
-                }
-
-                RightSidebar {
-                    lang,
-                    promotion: hot_promotions,
-                    news,
-                    followers,
-
-                    follow: move |id: i64| async move {
-                        ctrl.follow(id).await;
-                    },
-                }
-            }
-
-            div {
-                class: "fixed bottom-85 left-0 w-full hidden max-tablet:flex aria-hidden:!hidden",
-                aria_hidden: is_write(),
-                BottomSheet {
-                    lang,
-                    profile: profile.clone(),
-                    recent_feeds,
-                    recent_spaces,
-                    recent_communities,
-                    accounts,
-
-                    add_account: move |_| async move {
-                        ctrl.add_account().await;
-                    },
-                    sign_out: move |_| async move {
-                        ctrl.signout().await;
-                    },
-                }
-            }
-
-            div {
-                class: "flex flex-row w-full justify-start items-start aria-hidden:!hidden",
-                aria_hidden: !is_write(),
-                CreateFeedBox {
-                    lang,
-                    nickname: profile.nickname.clone(),
-                    profile: profile.profile.clone(),
-                    onclose: move |_| {
-                        is_write.set(false);
-                    },
-                    onsend: move |(content_type, description): (ContentType, String)| async move {
-                        ctrl.create_feed(content_type, description).await;
-                    },
-                }
-            }
+            is_write: is_write(),
+            onwrite: move |_| {
+                is_write.set(true);
+            },
+            onclick: move |id: i64| {
+                ctrl.move_to_threads(id);
+            },
         }
     }
 }
